+++ dart/gui/osg/Utils.hpp
@@ -38,32 +38,86 @@
 #include <osg/Matrix>
 
 //==============================================================================
-template<typename Scalar>
-::osg::Matrix eigToOsgMatrix(const Eigen::Transform<Scalar,3,Eigen::Isometry>& tf)
+template <typename Scalar>
+::osg::Matrix eigToOsgMatrix(
+    const Eigen::Transform<Scalar, 3, Eigen::Isometry>& tf)
 {
   return ::osg::Matrix(
-        tf(0, 0), tf(1, 0), tf(2, 0), tf(3, 0),
-        tf(0, 1), tf(1, 1), tf(2, 1), tf(3, 1),
-        tf(0, 2), tf(1, 2), tf(2, 2), tf(3, 2),
-        tf(0, 3), tf(1, 3), tf(2, 3), tf(3, 3));
+      tf(0, 0),
+      tf(1, 0),
+      tf(2, 0),
+      tf(3, 0),
+      tf(0, 1),
+      tf(1, 1),
+      tf(2, 1),
+      tf(3, 1),
+      tf(0, 2),
+      tf(1, 2),
+      tf(2, 2),
+      tf(3, 2),
+      tf(0, 3),
+      tf(1, 3),
+      tf(2, 3),
+      tf(3, 3));
 }
 
 //==============================================================================
-template<typename Derived>
+template <typename Derived>
 ::osg::Matrix eigToOsgMatrix(const Eigen::DenseBase<Derived>& M)
 {
   return ::osg::Matrix(
-        M(0, 0), M(1, 0), M(2, 0), M(3, 0),
-        M(0, 1), M(1, 1), M(2, 1), M(3, 1),
-        M(0, 2), M(1, 2), M(2, 2), M(3, 2),
-        M(0, 3), M(1, 3), M(2, 3), M(3, 3));
+      M(0, 0),
+      M(1, 0),
+      M(2, 0),
+      M(3, 0),
+      M(0, 1),
+      M(1, 1),
+      M(2, 1),
+      M(3, 1),
+      M(0, 2),
+      M(1, 2),
+      M(2, 2),
+      M(3, 2),
+      M(0, 3),
+      M(1, 3),
+      M(2, 3),
+      M(3, 3));
 }
 
 //==============================================================================
-template<typename Derived>
-::osg::Vec3d eigToOsgVec3(const Eigen::MatrixBase<Derived>& vec)
+template <typename Derived>
+::osg::Vec3f eigToOsgVec3f(const Eigen::MatrixBase<Derived>& vec)
+{
+  return ::osg::Vec3f(vec[0], vec[1], vec[2]);
+}
+
+//==============================================================================
+template <typename Derived>
+::osg::Vec3d eigToOsgVec3d(const Eigen::MatrixBase<Derived>& vec)
 {
   return ::osg::Vec3d(vec[0], vec[1], vec[2]);
+}
+
+//==============================================================================
+template <typename Derived>
+typename std::conditional<
+    std::is_same<typename Derived::Scalar, float>::value,
+    ::osg::Vec3f,
+    ::osg::Vec3d>::type
+eigToOsgVec3(const Eigen::MatrixBase<Derived>& vec)
+{
+  using Vec3 = typename std::conditional<
+      std::is_same<typename Derived::Scalar, float>::value,
+      ::osg::Vec3f,
+      ::osg::Vec3d>::type;
+
+  return Vec3(vec[0], vec[1], vec[2]);
+}
+
+//==============================================================================
+inline Eigen::Vector3f osgToEigVec3(const ::osg::Vec3f& vec)
+{
+  return Eigen::Vector3f(vec[0], vec[1], vec[2]);
 }
 
 //==============================================================================
@@ -73,10 +127,37 @@ inline Eigen::Vector3d osgToEigVec3(const ::osg::Vec3d& vec)
 }
 
 //==============================================================================
-template<typename Derived>
-::osg::Vec4d eigToOsgVec4(const Eigen::MatrixBase<Derived>& vec)
+template <typename Derived>
+::osg::Vec4f eigToOsgVec4f(const Eigen::MatrixBase<Derived>& vec)
+{
+  return ::osg::Vec4f(vec[0], vec[1], vec[2], vec[3]);
+}
+
+//==============================================================================
+template <typename Derived>
+::osg::Vec4d eigToOsgVec4d(const Eigen::MatrixBase<Derived>& vec)
 {
   return ::osg::Vec4d(vec[0], vec[1], vec[2], vec[3]);
+}
+
+//==============================================================================
+template <typename Derived>
+std::conditional<
+    std::is_same<typename Derived::Scalar, float>::value,
+    ::osg::Vec4f,
+    ::osg::Vec4d>
+eigToOsgVec4(const Eigen::MatrixBase<Derived>& vec)
+{
+  return std::conditional<
+      std::is_same<typename Derived::Scalar, float>::value,
+      ::osg::Vec4f,
+      ::osg::Vec4d>(vec[0], vec[1], vec[2], vec[3]);
+}
+
+//==============================================================================
+inline Eigen::Vector4f osgToEigVec4(const ::osg::Vec4f& vec)
+{
+  return Eigen::Vector4f(vec[0], vec[1], vec[2], vec[3]);
 }
 
 //==============================================================================