+++ dart/gui/osg/render/CapsuleShapeNode.cpp
@@ -205,7 +205,7 @@ void CapsuleShapeDrawable::refresh(bool firstTime)
   if(mCapsuleShape->checkDataVariance(dart::dynamics::Shape::DYNAMIC_COLOR)
      || firstTime)
   {
-    setColor(eigToOsgVec4(mVisualAspect->getRGBA()));
+    setColor(eigToOsgVec4d(mVisualAspect->getRGBA()));
   }
 }