+++ dart/gui/osg/render/CylinderShapeNode.cpp
@@ -204,7 +204,7 @@ void CylinderShapeDrawable::refresh(bool firstTime)
   if(mCylinderShape->checkDataVariance(dart::dynamics::Shape::DYNAMIC_COLOR)
      || firstTime)
   {
-    setColor(eigToOsgVec4(mVisualAspect->getRGBA()));
+    setColor(eigToOsgVec4d(mVisualAspect->getRGBA()));
   }
 }