+++ dart/gui/osg/render/EllipsoidShapeNode.cpp
@@ -224,7 +224,7 @@ void EllipsoidShapeDrawable::refresh(bool firstTime)
   if(mEllipsoidShape->checkDataVariance(dart::dynamics::Shape::DYNAMIC_COLOR)
      || firstTime)
   {
-    setColor(eigToOsgVec4(mVisualAspect->getRGBA()));
+    setColor(eigToOsgVec4d(mVisualAspect->getRGBA()));
   }
 }