+++ dart/gui/osg/render/PointCloudShapeNode.cpp
@@ -222,7 +222,7 @@ void PointCloudShapeDrawable::refresh(bool /*firstTime*/)
   // the color could be updated in the version up. So we always update the
   // color.
   {
-    setColor(eigToOsgVec4(mVisualAspect->getRGBA()));
+    setColor(eigToOsgVec4d(mVisualAspect->getRGBA()));
   }
 }